+++ resources/frontend/medicardice/src/app/secure/user/user.component.ts
@@ -1,22 +1,29 @@
-import {Component, OnInit} from '@angular/core';
+import {Component, Inject, OnInit, PLATFORM_ID, signal} from '@angular/core';
 import {FormBuilder, FormGroup, Validators} from "@angular/forms";
 import {UserService} from "../../shared/services/user/user.service";
-import {User} from "../../shared/models/user.response.login";
+import {Role, User} from "../../shared/models/user.response.login";
 import {ActivatedRoute, Router} from "@angular/router";
 import {AuthService} from "../../public/auth.service";
 import {ExpiresAtService} from "../../shared/services/expires-at.service";
 import {first} from "rxjs";
+import {isPlatformBrowser} from "@angular/common";
+import {FormHeaderComponent} from "../components/form-header/form-header.component";
 
 @Component({
   selector: 'app-user',
   templateUrl: './user.component.html',
-  styleUrl: './user.component.css'
+  styleUrl: './user.component.css',
+
 })
 export class UserComponent implements OnInit{
   frmGroupUser!:FormGroup;
   insertedUser!:User;
   //userData!:UserData;
   name!:string;
+  titre:string='ajout';
+  childTitle=signal('administrateur')
+  label=signal('liste');
+  titreButton:string='ajouter';
   isExpired!:boolean;
   id!:number;
   isAddMode!:boolean;
@@ -27,14 +34,16 @@ export class UserComponent implements OnInit{
     private router:Router,
     private route:ActivatedRoute,
     private authService:AuthService,
-    private expireService:ExpiresAtService
+    private expireService:ExpiresAtService,
+    @Inject(PLATFORM_ID) private platformId: Object
   ) {
       this.frmGroupUser=this.fb.group({
         last_name:['',[Validators.required]],
         first_name:['',[Validators.required]],
         email:['',[Validators.required]],
-        password:['',[Validators.required]],
-        password_confirm:['',[Validators.required]],
+        role:[''],
+        password:['123456'],
+        password_confirm:['123456'],
       })
   }
   get last_name(){
@@ -52,32 +61,50 @@ export class UserComponent implements OnInit{
   get password_confirm(){
     return this.frmGroupUser.get('password_confirm')
   }
+  get role(){
+    return this.frmGroupUser.get('role')
+  }
   onSubmit() {
-    this.userService.create(this.frmGroupUser.value)
-      .subscribe({
-        next:user=>{
-          this.insertedUser=user
-          if(user.success){
-            const fullName=`${this.insertedUser.data.last_name} ${this.insertedUser.data.first_name}`
-            this.router.navigate(['activation',fullName])
-          }
-        },
-        error:err =>console.log(err.error)
-      });
+    //console.log(this.frmGroupUser.value)
+    if(this.isAddMode)
+    {
+      this.userService.create(this.frmGroupUser.value)
+        .subscribe({
+          next:user=>{
+            this.insertedUser=user
+            this.router.navigateByUrl('dashboard/user/list')
+            /*if(user.success){
+              const fullName=`${this.insertedUser.data.last_name} ${this.insertedUser.data.first_name}`
+              this.router.navigate(['activation',fullName])
+            }*/
+          },
+          error:err =>console.log(err.error)
+        });
+    }
+    else{
+      this.userService.update(this.id,this.frmGroupUser.value)
+        .subscribe({
+          next:res=>{
+            if(res.success){
+              this.router.navigateByUrl('dashboard/user/list')
+            }
+          },
+          error:err=>console.log(err)
+        })
+    }
+
   }
 
   ngOnInit(): void {
-
       this.expireService.updateState(this.authService.isExpired());
       this.expireService.state$.subscribe(res=>this.isExpired=res);
       if(this.isExpired){
-        /*const token =`Bearer `+localStorage.getItem('token');
-        this.authService.logout(token)
-          .subscribe(res=>console.log(res))*/
-        localStorage.clear();
+        if(isPlatformBrowser(this.platformId)){
+          localStorage.clear();
+        }
+
         this.router.navigateByUrl('login',{replaceUrl:true})
           .then(()=>{
-
             this.router.navigate([this.router.url])
           })
       }
@@ -85,20 +112,27 @@ export class UserComponent implements OnInit{
       this.id=this.route.snapshot.params['id'];
       this.isAddMode=!this.id
       if(!this.isAddMode){
+        this.titre='mise à jour'
         this.userService.show(this.id)
           .pipe(first())
           .subscribe({
             next:(res)=>{
-              //const key:string|any='nom'
               const dataKey:number|any=0
               this.insertedUser=res
+              //console.log(this.insertedUser)
+              // @ts-ignore
+              const roles =this.insertedUser.data[dataKey].roles as Role
+              // @ts-ignore
+
               this.frmGroupUser.patchValue({
                 // @ts-ignore
                 email:this.insertedUser.data[dataKey].email,
                 // @ts-ignore
                 last_name:this.insertedUser.data[dataKey].nom,
                 // @ts-ignore
-                first_name:this.insertedUser.data[dataKey].prenom
+                first_name:this.insertedUser.data[dataKey].prenom,
+                // @ts-ignore
+                role:roles[0].role
               });
             },
             error:(err)=>console.log(err)
